simplify Parameters Build/Plasma
