+++ src/actors/balance_plant/thermal_plant_actor.jl
@@ -8,7 +8,7 @@ import ThermalSystemModels
 TSMD = ThermalSystemModels.Dynamics
 MTK.@variables t
 
-Base.@kwdef mutable struct FUSEparameters__ActorThermalPlant{T<:Real} <: ParametersActorBuild{T}
+Base.@kwdef mutable struct FUSEparameters__ActorThermalPlant{T<:Real} <: ParametersActor{T}
     _parent::WeakRef = WeakRef(Nothing)
     _name::Symbol = :not_set
     _time::Float64 = NaN
