+++ src/actors/costing/costing_actor.jl
@@ -4,7 +4,7 @@ using DataFrames
 #= ============ =#
 #  ActorCosting  #
 #= ============ =#
-Base.@kwdef mutable struct FUSEparameters__ActorCosting{T<:Real} <: ParametersActorBuild{T}
+Base.@kwdef mutable struct FUSEparameters__ActorCosting{T<:Real} <: ParametersActor{T}
     _parent::WeakRef = WeakRef(nothing)
     _name::Symbol = :not_set
     _time::Float64 = NaN
