+++ src/actors/costing/sheffield_costing_actor.jl
@@ -2,7 +2,7 @@
 #  ActorCostingSheffield  #
 #= ===================== =#
 
-Base.@kwdef mutable struct FUSEparameters__ActorCostingSheffield{T<:Real} <: ParametersActorBuild{T}
+Base.@kwdef mutable struct FUSEparameters__ActorCostingSheffield{T<:Real} <: ParametersActor{T}
     _parent::WeakRef = WeakRef(nothing)
     _name::Symbol = :not_set
     _time::Float64 = NaN
