+++ src/actors/current/current_actor.jl
@@ -1,7 +1,7 @@
 #= ============ =#
 #  ActorCurrent  #
 #= ============ =#
-Base.@kwdef mutable struct FUSEparameters__ActorCurrent{T<:Real} <: ParametersActorPlasma{T}
+Base.@kwdef mutable struct FUSEparameters__ActorCurrent{T<:Real} <: ParametersActor{T}
     _parent::WeakRef = WeakRef(nothing)
     _name::Symbol = :not_set
     _time::Float64 = NaN
