+++ src/actors/hcd/ec_simple_actor.jl
@@ -1,7 +1,7 @@
 #= == =#
 #  EC  #
 #= == =#
-Base.@kwdef mutable struct FUSEparameters__ActorSimpleEC{T<:Real} <: ParametersActorPlasma{T}
+Base.@kwdef mutable struct FUSEparameters__ActorSimpleEC{T<:Real} <: ParametersActor{T}
     _parent::WeakRef = WeakRef(nothing)
     _name::Symbol = :not_set
     _time::Float64 = NaN
