+++ src/actors/hcd/ic_simple_actor.jl
@@ -1,7 +1,7 @@
 #= == =#
 #  IC  #
 #= == =#
-Base.@kwdef mutable struct FUSEparameters__ActorSimpleIC{T<:Real} <: ParametersActorPlasma{T}
+Base.@kwdef mutable struct FUSEparameters__ActorSimpleIC{T<:Real} <: ParametersActor{T}
     _parent::WeakRef = WeakRef(nothing)
     _name::Symbol = :not_set
     _time::Float64 = NaN
