+++ src/actors/hcd/lh_simple_actor.jl
@@ -1,7 +1,7 @@
 #= == =#
 #  LH  #
 #= == =#
-Base.@kwdef mutable struct FUSEparameters__ActorSimpleLH{T<:Real} <: ParametersActorPlasma{T}
+Base.@kwdef mutable struct FUSEparameters__ActorSimpleLH{T<:Real} <: ParametersActor{T}
     _parent::WeakRef = WeakRef(nothing)
     _name::Symbol = :not_set
     _time::Float64 = NaN
