+++ src/actors/hcd/nbi/nb_simple_actor.jl
@@ -1,7 +1,7 @@
 #= ========== =#
 #  Simple NBI  #
 #= ========== =#
-Base.@kwdef mutable struct FUSEparameters__ActorSimpleNB{T<:Real} <: ParametersActorPlasma{T}
+Base.@kwdef mutable struct FUSEparameters__ActorSimpleNB{T<:Real} <: ParametersActor{T}
     _parent::WeakRef = WeakRef(nothing)
     _name::Symbol = :not_set
     _time::Float64 = NaN
