+++ src/actors/hcd/nbi/nbi_actor.jl
@@ -1,7 +1,7 @@
 #= === =#
 #  NBI  #
 #= === =#
-Base.@kwdef mutable struct FUSEparameters__ActorNBI{T<:Real} <: ParametersActorPlasma{T}
+Base.@kwdef mutable struct FUSEparameters__ActorNBI{T<:Real} <: ParametersActor{T}
     _parent::WeakRef = WeakRef(nothing)
     _name::Symbol = :not_set
     _time::Float64 = NaN
