+++ src/actors/hcd/nbi/rabbit_actor.jl
@@ -4,7 +4,7 @@ using Plots
 #= =========== =#
 #  ActorRABBIT  #
 #= =========== =#
-Base.@kwdef mutable struct FUSEparameters__ActorRABBIT{T<:Real} <: ParametersActorPlasma{T}
+Base.@kwdef mutable struct FUSEparameters__ActorRABBIT{T<:Real} <: ParametersActor{T}
     _parent::WeakRef = WeakRef(nothing)
     _name::Symbol = :not_set
     _time::Float64 = NaN
