+++ src/actors/hcd/pellet_simple_actor.jl
@@ -1,7 +1,7 @@
 #= ====== =#
 #  PELLET  #
 #= ====== =#
-Base.@kwdef mutable struct FUSEparameters__ActorSimplePellet{T<:Real} <: ParametersActorPlasma{T}
+Base.@kwdef mutable struct FUSEparameters__ActorSimplePellet{T<:Real} <: ParametersActor{T}
     _parent::WeakRef = WeakRef(nothing)
     _name::Symbol = :not_set
     _time::Float64 = NaN
