+++ src/actors/noop_actor.jl
@@ -1,7 +1,7 @@
 #= ================ =#
 #  ActorNoOperation  #
 #= ================ =#
-Base.@kwdef mutable struct FUSEparameters__ActorNoOperation{T<:Real} <: ParametersActorPlasma{T}
+Base.@kwdef mutable struct FUSEparameters__ActorNoOperation{T<:Real} <: ParametersActor{T}
     _parent::WeakRef = WeakRef(nothing)
     _name::Symbol = :not_set
     _time::Float64 = NaN
