+++ src/actors/nuclear/blanket_actor.jl
@@ -3,7 +3,7 @@ import NNeutronics
 #= ============ =#
 #  ActorBlanket  #
 #= ============ =#
-Base.@kwdef mutable struct FUSEparameters__ActorBlanket{T<:Real} <: ParametersActorBuild{T}
+Base.@kwdef mutable struct FUSEparameters__ActorBlanket{T<:Real} <: ParametersActor{T}
     _parent::WeakRef = WeakRef(nothing)
     _name::Symbol = :not_set
     _time::Float64 = NaN
