+++ src/actors/pedestal/EPED_actor.jl
@@ -3,7 +3,7 @@ import EPEDNN
 #= ========= =#
 #  ActorEPED  #
 #= ========= =#
-Base.@kwdef mutable struct FUSEparameters__ActorEPED{T<:Real} <: ParametersActorPlasma{T}
+Base.@kwdef mutable struct FUSEparameters__ActorEPED{T<:Real} <: ParametersActor{T}
     _parent::WeakRef = WeakRef(nothing)
     _name::Symbol = :not_set
     _time::Float64 = NaN
