+++ src/actors/pedestal/WPED_actor.jl
@@ -1,7 +1,7 @@
 #= ========= =#
 #  ActorWPED  #
 #= ========= =#
-Base.@kwdef mutable struct FUSEparameters__ActorWPED{T<:Real} <: ParametersActorPlasma{T}
+Base.@kwdef mutable struct FUSEparameters__ActorWPED{T<:Real} <: ParametersActor{T}
     _parent::WeakRef = WeakRef(nothing)
     _name::Symbol = :not_set
     _time::Float64 = NaN
