+++ src/actors/pedestal/pedestal_actor.jl
@@ -3,7 +3,7 @@ import EPEDNN
 #= ============= =#
 #  ActorPedestal  #
 #= ============= =#
-Base.@kwdef mutable struct FUSEparameters__ActorPedestal{T<:Real} <: ParametersActorPlasma{T}
+Base.@kwdef mutable struct FUSEparameters__ActorPedestal{T<:Real} <: ParametersActor{T}
     _parent::WeakRef = WeakRef(nothing)
     _name::Symbol = :not_set
     _time::Float64 = NaN
