+++ src/actors/pf/pf_active_actor.jl
@@ -6,7 +6,7 @@ import VacuumFields
 #= ============= =#
 #  ActorPFactive  #
 #= ============= =#
-Base.@kwdef mutable struct FUSEparameters__ActorPFactive{T<:Real} <: ParametersActorPlasma{T}
+Base.@kwdef mutable struct FUSEparameters__ActorPFactive{T<:Real} <: ParametersActor{T}
     _parent::WeakRef = WeakRef(nothing)
     _name::Symbol = :not_set
     _time::Float64 = NaN
