+++ src/actors/stability/limits_actor.jl
@@ -1,7 +1,7 @@
 #= ==================== =#
 #  ActorStabilityLimits  #
 #= ==================== =#
-Base.@kwdef mutable struct FUSEparameters__ActorStabilityLimits{T<:Real} <: ParametersActorPlasma{T}
+Base.@kwdef mutable struct FUSEparameters__ActorStabilityLimits{T<:Real} <: ParametersActor{T}
     _parent::WeakRef = WeakRef(nothing)
     _name::Symbol = :not_set
     _time::Float64 = NaN
