+++ src/actors/transport/core_transport_actor.jl
@@ -1,7 +1,7 @@
 #= ================== =#
 #  ActorCoreTransport  #
 #= ================== =#
-Base.@kwdef mutable struct FUSEparameters__ActorCoreTransport{T<:Real} <: ParametersActorPlasma{T}
+Base.@kwdef mutable struct FUSEparameters__ActorCoreTransport{T<:Real} <: ParametersActor{T}
     _parent::WeakRef = WeakRef(nothing)
     _name::Symbol = :not_set
     _time::Float64 = NaN
