+++ src/actors/transport/eped_profiles_actor.jl
@@ -1,7 +1,7 @@
 #= ================== =#
 #  ActorEPEDprofiles  #
 #= ================== =#
-Base.@kwdef mutable struct FUSEparameters__ActorEPEDprofiles{T<:Real} <: ParametersActorPlasma{T}
+Base.@kwdef mutable struct FUSEparameters__ActorEPEDprofiles{T<:Real} <: ParametersActor{T}
     _parent::WeakRef = WeakRef(nothing)
     _name::Symbol = :not_set
     _time::Float64 = NaN
