+++ src/actors/transport/flux_calculator_actor.jl
@@ -1,7 +1,7 @@
 #= =================== =#
 #  ActorFluxCalculator  #
 #= =================== =#
-Base.@kwdef mutable struct FUSEparameters__ActorFluxCalculator{T<:Real} <: ParametersActorPlasma{T}
+Base.@kwdef mutable struct FUSEparameters__ActorFluxCalculator{T<:Real} <: ParametersActor{T}
     _parent::WeakRef = WeakRef(nothing)
     _name::Symbol = :not_set
     _time::Float64 = NaN
