+++ src/actors/transport/neoclassical_actor.jl
@@ -3,7 +3,7 @@ import NEO
 #= ================= =#
 #  ActorNeoclassical  #
 #= ================= =#
-Base.@kwdef mutable struct FUSEparameters__ActorNeoclassical{T<:Real} <: ParametersActorPlasma{T}
+Base.@kwdef mutable struct FUSEparameters__ActorNeoclassical{T<:Real} <: ParametersActor{T}
     _parent::WeakRef = WeakRef(nothing)
     _name::Symbol = :not_set
     _time::Float64 = NaN
