+++ src/actors/transport/qlgyro_actor.jl
@@ -4,7 +4,7 @@ import TGLFNN: InputQLGYRO, InputCGYRO
 #= =========== =#
 #  ActorQLGYRO  #
 #= =========== =#
-Base.@kwdef mutable struct FUSEparameters__ActorQLGYRO{T<:Real} <: ParametersActorPlasma{T}
+Base.@kwdef mutable struct FUSEparameters__ActorQLGYRO{T<:Real} <: ParametersActor{T}
     _parent::WeakRef = WeakRef(nothing)
     _name::Symbol = :not_set
     _time::Float64 = NaN
