+++ src/actors/wall_loading/corerad_hf_actor.jl
@@ -2,7 +2,7 @@
 #  ActorCoreRadHeatFlux  #
 #= ===================== =#
 
-Base.@kwdef mutable struct FUSEparameters__ActorCoreRadHeatFlux{T<:Real} <: ParametersActorPlasma{T}
+Base.@kwdef mutable struct FUSEparameters__ActorCoreRadHeatFlux{T<:Real} <: ParametersActor{T}
     _parent::WeakRef = WeakRef(nothing)
     _name::Symbol = :not_set
     _time::Float64 = NaN
