+++ src/actors/wall_loading/particle_hf_actor.jl
@@ -2,7 +2,7 @@
 #  ActorParticleHeatFlux  #
 #= ===================== =#
 
-Base.@kwdef mutable struct FUSEparameters__ActorParticleHeatFlux{T<:Real} <: ParametersActorPlasma{T}
+Base.@kwdef mutable struct FUSEparameters__ActorParticleHeatFlux{T<:Real} <: ParametersActor{T}
     _parent::WeakRef = WeakRef(nothing)
     _name::Symbol = :not_set
     _time::Float64 = NaN
