+++ src/parameters.jl
@@ -11,13 +11,9 @@ function SimulationParameters.Switch{T}(options, ids::Type{<:IMAS.IDS}, field::S
 end
 
 abstract type ParametersActor{T} <: AbstractParameters{T} end # abstract type for parameters of an actor (each actor has its concrete type)
-abstract type ParametersActorPlasma{T} <: ParametersActor{T} end
-abstract type ParametersActorBuild{T} <: ParametersActor{T} end
 abstract type ParametersAllActors{T} <: AbstractParameters{T} end # --> abstract type for all parameters of all actors (the concrete type of this is ParametersActors)
 
 abstract type ParametersInit{T} <: AbstractParameters{T} end # abstract type for parameters of a init (each init has its concrete type)
-abstract type ParametersInitPlasma{T} <: ParametersInit{T} end
-abstract type ParametersInitBuild{T} <: ParametersInit{T} end
 abstract type ParametersAllInits{T} <: AbstractParameters{T} end # --> abstract type for all parameters of all inits (the concrete type of this is ParametersInits)
 
 function SimulationParameters.time_range(parameters::ParametersInit)
